bug fixes for converters handling unsigned integers, chars, etc.
The RTTR example these were derived from handled char as a bool,
which caused data loss.  That is now fixed.

JsonGLIB stores all integers as 64-bit itegers; the library
is used by both the jsonglib and socket-io converters, and the
converters were both written with the incorrect assumption that
these types would be internally handled okay.  Instead the use
of static casting combined with checking the registered RTTR
data type are required for ensuring the value is restored in
its original form.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>