meson: make devenv PATH include rttr
CMake subprojects, or perhaps something bugged with rttr,
results in Meson <= 1.0.1 failing to put the builddir on
the PATH variable when using 'meson devenv -C builddir'
if you are also using one of the VS backends.  Because
of this, any use of the VS solution requires the user
to update the solution's environment XML manually to include
the library, else one encounters the library not found
error.  By setting it in the devenv this way, one can
go straight from 'meson devenv' to running VS from that
shell and have all the required paths configured.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>