metadata: fixed 'check' of blob, no-serialize bool flags
The original check only validated that the flag was valid,
not its actual value.  This more-correctly checks the value
that was stored and returns it.  There was no actual bug
in the original code since we only store 'true' if storing
the flag at all, so .is_valid() is okay, however if we ever
change that behavior to store something enumerated, etc.
then this commit will help flag the change in behavior if
the change cannot be coerced into a boolean.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>