converters: fix for handling pointer and shared_ptr
When coming "from" on pointer-types, we have to call the
constructor.  For shared pointers (and ref, etc.), these
are "wrapped" around a pointer, so first one has to check
for if the RTTR Type is "wrapped" and if so, get the inner
wrapped type.  Then, if that type is a pointer, which it
likely will be, create an instance using a constructor
that matches the original type.  This means the type that
has declared a pointer-type member needs to provide a
constructor registration on that object's definition
or refer to a static method (again, via registration) that
can produce the required type.  The loops shown in this commit
will try to find the matching constructor and call that.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>