converters: added support for std::any in map values
See the associated test for details.  Use with caution,
there are drawbacks since the intermediate type (JSON, etc.)
may not support all flavors of integer, so on restoration of
that type it becomes a std::any containing the value type
returned by the associated API used to retrieve that member
(in that case, a 64-bit number rather than perhaps a smaller
type).

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>