added rttr patch to fix Windows.h compatibility
Windows.h (ultimately) defines macros for min and max, which
the way these files are written causes them to break if that
header is included first (which is sometimes unavoidable).
Wrapping the call to max with parenthesis disables the macro
from being applied in these places, thus allowing the STL
versions to be used.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>