bug fix: optional-with-default empty-like member behavior
Prior to this patch, an empty-like member, e.g., std::string,
that had optiona-with-default set to a reference value would
be skipped if it is either empty or matched the string.  With
this patch, if a default is provided, the matching of that
default value will determine if the member will revert to
being treated as required.  In other words, matching the
provided default means it can be treated as actually optional
(based on its value) and failing to match that reference means
its value is now required.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>